Fix Buffer Size
The buffer size in the Bmode image was to large. The buffer size is unaffacted
by scanWidth and scanHeight in the ScanConverter and appers to have a fixed size
of NBOLINES x MAX_SAMPLES.

Now teh BMode images is received at full depth and the image size does not ahve to
be set to MAX_SMAPLES/2.