ENH: Update for Interson SDK version 2.21
I modified the value of IntersonArraySDK_DIR to point to the actual folder containing the DLLs, not a parent folder a couple of steps up.  The folder structure is different in the latest version of the SDK.

I changed the HWButtonTick (updated to new class names) handling to work with the new/replacement Release event in this version of the SDK.

Added IntersonArrayServerBMode.

Fixed typo in IntersonArrayServerRF.xml.