ENH: more robust preprocessing for Clarius and Butterfly
Clarius has reduced the letterboxing around their videos/images causing rulers to overlay the B-mode.  New method is robust to this but slower due to the inpainting.

Butterfly now adds a gray bar to the top of their videos, we now crop from the center and should be more robust.